feat: extract v06 indexer logic
