+++ user-ops-indexer/user-ops-indexer-entity/src/user_operations.rs
@@ -1,7 +1,6 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
-use super::sea_orm_active_enums::EntryPointVersion;
-use super::sea_orm_active_enums::SponsorType;
+use super::sea_orm_active_enums::{EntryPointVersion, SponsorType};
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
