+++ user-ops-indexer/user-ops-indexer-logic/src/indexer/mod.rs
@@ -1,1 +1,3 @@
+mod base_indexer;
 pub mod v06;
+pub use base_indexer::Indexer;
